Add types back into Select component
